+++ src/application/algos/dfs.rs
@@ -1,4 +1,5 @@
 use rayon::iter::ParallelIterator;
+use rust_decimal::Decimal;
 
 use crate::adapter::graph::Graph;
 use crate::domain::types::{ExecutionParams, Route, Side, SwapStep, Token};
@@ -22,7 +23,7 @@ pub fn dfs(
     graph: &Graph,
     from: &Token,
     to: &Token,
-    amount_in: f64,
+    amount_in: Decimal,
     params: ExecutionParams,
 ) -> Option<Route> {
     let mut stack = VecDeque::new();
@@ -33,7 +34,7 @@ pub fn dfs(
         HashSet::from([from.clone()]),
     ));
 
-    let mut visited: HashMap<Token, f64> = HashMap::new();
+    let mut visited: HashMap<Token, Decimal> = HashMap::new();
 
     while let Some((token, route, cumulative_amount, seen)) = stack.pop_back() {
         if token == *to {
